requirements.txt
# Book Metadata and Recommendation API - Dependencies
# Python 3.x required

# Core Framework
Django==6.0.3
djangorestframework==3.17.1

# API Documentation
drf-yasg==1.21.15

# Filtering and Search
django-filter==25.2

# Additional Dependencies (auto-installed with above)
asgiref>=3.9.1
sqlparse>=0.5.0
tzdata
pytz>=2021.1
pyyaml>=5.1
uritemplate>=3.0.0
inflection>=0.3.1
packaging>=21.0

# Development Tools (optional)
# requests  # For API testing
# pytest    # For unit testing
# pytest-django  # Django integration for pytest
